refactor(whatsrust): extract query models
